Add base support for 3d tiles layers (#7986)
