+++ web/client/api/catalog/index.js
@@ -7,14 +7,13 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import csw from '../CSW';
-import wms from '../WMS';
-import wmts from '../WMTS';
+import * as csw from './CSW';
+import * as wms from './WMS';
+import * as wmts from './WMTS';
 import * as tms from './TMS';
 import * as wfs from './WFS';
-import backgrounds from '../mapBackground';
-import { validate, testService, preprocess } from './common';
-
+import * as backgrounds from './backgrounds';
+import * as threeDTiles from './ThreeDTiles';
 
 /**
  * APIs collection for catalog.
@@ -30,7 +29,9 @@ import { validate, testService, preprocess } from './common';
  *      }
  * }
  * ```
- * Optionally implements validation functions:
+ * - `getCatalogRecords` (data, options) => function that returns an array of catalogs records
+ * - `getLayerFromRecord` (record, options, asPromise) => function that returns a promise/object that resolve with a mapstore layer configuration object given a catalog record
+ * - `preprocess` return an Observable that performs actions on service object prior to its save
  * - `validate`: function that gets the service object and returns an Observable. The stream emit an exception if the service validation fails. Otherwise it emits the `service` object and complete.
  * - `testService` function that gets the service object and returns an Observable. The stream emit an exception if the service do not respond. Otherwise it emits the `service` object and complete.
  * @memberof api
@@ -40,35 +41,11 @@ export default {
     // TODO: we should separate catalog specific API from OGC services API, to define better the real interfaces of each API.
     // TODO: validate could be converted in a simple function
     // TODO: testService could be converted in a simple Promise
-    csw: {
-        preprocess,
-        validate,
-        testService: testService(csw),
-        ...csw
-    },
-    wfs: {
-        preprocess,
-        validate,
-        testService: testService(wfs),
-        ...wfs
-    },
-    wms: {
-        preprocess,
-        validate,
-        testService: testService(wms),
-        ...wms
-    },
-    tms, // has it's own validation
-    wmts: {
-        preprocess,
-        validate,
-        testService: testService(wmts),
-        ...wmts
-    },
-    backgrounds: {
-        preprocess,
-        validate,
-        testService: testService(backgrounds),
-        ...backgrounds
-    }
+    'csw': csw,
+    'wfs': wfs,
+    'wms': wms,
+    'tms': tms,
+    'wmts': wmts,
+    'backgrounds': backgrounds,
+    '3dtiles': threeDTiles
 };
