+++ web/client/components/TOC/fragments/settings/Display.jsx
@@ -18,7 +18,7 @@ import InfoPopover from '../../../widgets/widget/InfoPopover';
 import Legend from '../legend/Legend';
 import VisibilityLimitsForm from './VisibilityLimitsForm';
 import Select from 'react-select';
-import { DEFAULT_FORMAT_WMS, getSupportedFormat } from '../../../../utils/CatalogUtils';
+import { DEFAULT_FORMAT_WMS, getSupportedFormat } from '../../../../api/WMS';
 export default class extends React.Component {
     static propTypes = {
         opacityText: PropTypes.node,
@@ -167,7 +167,7 @@ export default class extends React.Component {
                     </Col>
                 </Row>}
 
-                <Row>
+                {this.props.element.type !== "3dtiles" && <Row>
                     <Col xs={12}>
                         <FormGroup>
                             <ControlLabel>{this.props.opacityText} %</ControlLabel>
@@ -180,7 +180,7 @@ export default class extends React.Component {
                                 onChange={(val)=> this.onChange("opacity", val)}/>
                         </FormGroup>
                     </Col>
-                </Row>
+                </Row>}
 
                 <Row>
                     <Col xs={12}>
