+++ web/client/components/catalog/__tests__/RecordItem-test.jsx
@@ -17,6 +17,8 @@ import RecordItem from '../RecordItem.jsx';
 const SAMPLE_IMAGE = "data:image/gif;base64,R0lGODlhAQABAIAAAAAAAP///yH5BAEAAAAALAAAAAABAAEAAAIBRAA7";
 
 const sampleRecord = {
+    serviceType: 'wms',
+    isValid: true,
     identifier: "test-identifier",
     title: "sample title",
     tags: ["subject1", "subject2"],
@@ -36,10 +38,18 @@ const sampleRecord = {
         url: "http://wms.sample.service:80/geoserver/wms?SERVICE=WMS&",
         SRS: [],
         params: {name: "workspace:layername"}
-    }]
+    }],
+    ogcReferences: {
+        type: "OGC:WMS",
+        url: "http://wms.sample.service:80/geoserver/wms?SERVICE=WMS&",
+        SRS: [],
+        params: {name: "workspace:layername"}
+    }
 };
 
 const sampleRecord2 = {
+    serviceType: 'wms',
+    isValid: true,
     identifier: "test-identifier",
     title: "sample title",
     tags: ["subject1", "subject2"],
@@ -58,10 +68,18 @@ const sampleRecord2 = {
         url: "http://wms.sample.service:80/geoserver/wms?SERVICE=WMS&",
         SRS: ['EPSG:4326'],
         params: {name: "workspace:layername"}
-    }]
+    }],
+    ogcReferences: {
+        type: "OGC:WMS",
+        url: "http://wms.sample.service:80/geoserver/wms?SERVICE=WMS&",
+        SRS: ['EPSG:4326'],
+        params: {name: "workspace:layername"}
+    }
 };
 
 const sampleRecord3 = {
+    serviceType: 'wmts',
+    isValid: true,
     identifier: "test-identifier",
     title: "sample title",
     tags: ["subject1", "subject2"],
@@ -81,7 +99,13 @@ const sampleRecord3 = {
         url: "http://wms.sample.service:80/geoserver/gwc/service/wmts",
         SRS: ['EPSG:4326', 'EPSG:3857'],
         params: {name: "workspace:layername"}
-    }]
+    }],
+    ogcReferences: {
+        type: "OGC:WMTS",
+        url: "http://wms.sample.service:80/geoserver/gwc/service/wmts",
+        SRS: ['EPSG:4326', 'EPSG:3857'],
+        params: {name: "workspace:layername"}
+    }
 };
 
 const getCapRecord = assign({}, sampleRecord, {references: [{
@@ -98,6 +122,8 @@ const getCapRecord = assign({}, sampleRecord, {references: [{
 ]});
 
 const longDescriptioRecord = {
+    serviceType: 'wms',
+    isValid: true,
     identifier: "test-identifier",
     tags: ["subject1", "subject2"],
     description: "Sed ut perspiciatis unde omnis iste natus error sit voluptatem accusantium doloremque laudantium, totam rem aperiam eaque ipsa, quae ab illo inventore veritatis et quasi architecto beatae vitae dicta sunt, explicabo. Nemo enim ipsam voluptatem, quia voluptas sit, aspernatur aut odit aut fugit, sed quia consequuntur magni dolores eos, qui ratione voluptatem sequi nesciunt, neque porro quisquam est, qui dolorem ipsum, quia dolor sit, amet, consectetur, adipisci velit, sed quia non numquam eius modi tempora incidunt, ut labore et dolore magnam aliquam quaerat voluptatem. Ut enim ad minima veniam, quis nostrum exercitationem ullam corporis suscipit laboriosam, nisi ut aliquid ex ea commodi consequatur? Quis autem vel eum iure reprehenderit, qui in ea voluptate velit esse, quam nihil molestiae consequatur, vel illum, qui dolorem eum fugiat, quo voluptas nulla pariatur? [33] At vero eos et accusamus et iusto odio dignissimos ducimus, qui blanditiis praesentium voluptatum deleniti atque corrupti, quos dolores et quas molestias excepturi sint, obcaecati cupiditate non provident, similique sunt in culpa, qui officia deserunt mollitia animi, id est laborum et dolorum fuga. Et harum quidem rerum facilis est et expedita distinctio. Nam libero tempore, cum soluta nobis est eligendi optio, cumque nihil impedit, quo minus id, quod maxime placeat, facere possimus, omnis voluptas assumenda est, omnis dolor repellendus. Temporibus autem quibusdam et aut officiis debitis aut rerum necessitatibus saepe eveniet, ut et voluptates repudiandae sint et molestiae non recusandae. Itaque earum rerum hic tenetur a sapiente delectus, ut aut reiciendis voluptatibus maiores alias consequatur aut perferendis doloribus asperiores repellat",
@@ -111,10 +137,19 @@ const longDescriptioRecord = {
         url: "http://wms.sample.service:80/geoserver/wms?SERVICE=WMS&",
         SRS: [],
         params: {name: "workspace:layername"}
-    }]
+    }],
+    ogcReferences: {
+        type: "OGC:WMS",
+        url: "http://wms.sample.service:80/geoserver/wms?SERVICE=WMS&",
+        SRS: [],
+        params: {name: "workspace:layername"}
+    }
 };
 
 const esriRecord = {
+    serviceType: 'csw',
+    layerType: 'esri',
+    isValid: true,
     title: "Esri Title",
     description: "Atlantic Hurricanes 2000",
     identifier: "f4bed551-faa2-4e9c-9820-e623098ba526",
@@ -147,154 +182,145 @@ describe('This test for RecordItem', () => {
     // test DEFAULTS
     it('creates the component with defaults', () => {
         const item = ReactDOM.render(<RecordItem />, document.getElementById("container"));
-        expect(item).toExist();
+        expect(item).toBeTruthy();
 
         const itemDom = ReactDOM.findDOMNode(item);
-        expect(itemDom).toNotExist();
+        expect(itemDom).toBeFalsy();
     });
     // test data
     it('creates the component with data', () => {
         const item = ReactDOM.render(<RecordItem record={sampleRecord}/>, document.getElementById("container"));
-        expect(item).toExist();
+        expect(item).toBeTruthy();
 
         const itemDom = ReactDOM.findDOMNode(item);
-        expect(itemDom).toExist();
+        expect(itemDom).toBeTruthy();
     });
-    it('check WMTS resource', () => {
-        let actions = {
-            onLayerAdd: () => {
-
-            }
-
-        };
-        let actionsSpy = expect.spyOn(actions, "onLayerAdd");
+    it('check WMTS resource', (done) => {
         const item = ReactDOM.render(<RecordItem
             record={sampleRecord3}
-            onLayerAdd={actions.onLayerAdd}
+            onLayerAdd={(layer, options) => {
+                expect(layer.format).toBe("image/png");
+                expect(options.zoomToLayer).toBeTruthy();
+                done();
+            }}
         />, document.getElementById("container"));
-        expect(item).toExist();
+        expect(item).toBeTruthy();
 
         const itemDom = ReactDOM.findDOMNode(item);
-        expect(itemDom).toExist();
+        expect(itemDom).toBeTruthy();
         let button = TestUtils.findRenderedDOMComponentWithTag(
             item, 'button'
         );
-        expect(button).toExist();
+        expect(button).toBeTruthy();
         button.click();
-        expect(actionsSpy.calls.length).toBe(1);
-        expect(actionsSpy.calls[0].arguments[0].format).toBe("image/png");
-        expect(actionsSpy.calls[0].arguments[1].zoomToLayer).toBeTruthy();
     });
-    it('check WMTS resource format', () => {
+    it('check WMTS resource format', (done) => {
         let actions = {
-            onLayerAdd: () => {
-
+            onLayerAdd: (layer, options) => {
+                expect(layer.format).toBe("image/jpeg");
+                expect(options.zoomToLayer).toBeTruthy();
+                done();
             }
         };
-        let actionsSpy = expect.spyOn(actions, "onLayerAdd");
         const item = ReactDOM.render(<RecordItem
             record={{...sampleRecord3, format: "image/jpeg"}}
             onLayerAdd={actions.onLayerAdd}
         />, document.getElementById("container"));
-        expect(item).toExist();
+        expect(item).toBeTruthy();
 
         const itemDom = ReactDOM.findDOMNode(item);
-        expect(itemDom).toExist();
+        expect(itemDom).toBeTruthy();
         let button = TestUtils.findRenderedDOMComponentWithTag(
             item, 'button'
         );
-        expect(button).toExist();
+        expect(button).toBeTruthy();
         button.click();
-        expect(actionsSpy.calls.length).toBe(1);
-        expect(actionsSpy.calls[0].arguments.length).toBe(2);
-        expect(actionsSpy.calls[0].arguments[0].format).toBe("image/jpeg");
-        expect(actionsSpy.calls[0].arguments[1].zoomToLayer).toBeTruthy();
     });
-    it('check esri resource', () => {
+    it('check esri resource', (done) => {
         let actions = {
-            onLayerAdd: () => {
-
+            onLayerAdd: (layer) => {
+                expect(layer).toBeTruthy();
+                done();
             }
         };
-        let actionsSpy = expect.spyOn(actions, "onLayerAdd");
-
         const item = ReactDOM.render(<RecordItem
             record={esriRecord}
             onLayerAdd={actions.onLayerAdd}/>, document.getElementById("container"));
-        expect(item).toExist();
+        expect(item).toBeTruthy();
 
         const itemDom = ReactDOM.findDOMNode(item);
-        expect(itemDom).toExist();
+        expect(itemDom).toBeTruthy();
         let button = TestUtils.findRenderedDOMComponentWithTag(
             item, 'button'
         );
-        expect(button).toExist();
+        expect(button).toBeTruthy();
         button.click();
-        expect(actionsSpy.calls.length).toBe(1);
     });
 
 
     // test handlers
-    it('check event handlers', () => {
+    it('check event handlers', (done) => {
         let actions = {
-            onLayerAdd: () => {
-
+            onLayerAdd: (layer, options) => {
+                expect(options.zoomToLayer).toBeTruthy();
+                done();
             }
         };
-        let actionsSpy = expect.spyOn(actions, "onLayerAdd");
         const item = ReactDOM.render(<RecordItem
             record={sampleRecord}
             onLayerAdd={actions.onLayerAdd}
         />, document.getElementById("container"));
-        expect(item).toExist();
+        expect(item).toBeTruthy();
 
         const itemDom = ReactDOM.findDOMNode(item);
-        expect(itemDom).toExist();
+        expect(itemDom).toBeTruthy();
         let button = TestUtils.findRenderedDOMComponentWithTag(
             item, 'button'
         );
-        expect(button).toExist();
+        expect(button).toBeTruthy();
         button.click();
-        expect(actionsSpy.calls.length).toBe(1);
-        expect(actionsSpy.calls[0].arguments[1].zoomToLayer).toBeTruthy();
     });
 
-    it('check event handlers with catalogUrl and csw service', () => {
+    it('check event handlers with catalogUrl and csw service', (done) => {
         let actions = {
-            onLayerAdd: () => {
-
+            onLayerAdd: (layer, options) => {
+                expect(layer.catalogURL).toBeTruthy();
+                expect(options.zoomToLayer).toBeTruthy();
+                done();
             }
         };
-        let actionsSpy = expect.spyOn(actions, "onLayerAdd");
         const item = ReactDOM.render(<RecordItem
             record={sampleRecord}
             onLayerAdd={actions.onLayerAdd}
             catalogURL="fakeURL"
             catalogType="csw"
         />, document.getElementById("container"));
-        expect(item).toExist();
+        expect(item).toBeTruthy();
 
         const itemDom = ReactDOM.findDOMNode(item);
-        expect(itemDom).toExist();
+        expect(itemDom).toBeTruthy();
         let button = TestUtils.findRenderedDOMComponentWithTag(
             item, 'button'
         );
-        expect(button).toExist();
+        expect(button).toBeTruthy();
         button.click();
-        expect(actionsSpy.calls.length).toBe(1);
-        expect(actionsSpy.calls[0].arguments.length).toBe(2);
-        expect(actionsSpy.calls[0].arguments[0].catalogURL).toExist();
-        expect(actionsSpy.calls[0].arguments[1].zoomToLayer).toBeTruthy();
     });
 
-    it('check event handlers with catalogUrl and wms service', () => {
+    it('check event handlers with catalogUrl and wms service', (done) => {
+        const service = { format: "image/jpeg", title: "GeoServer WMS", type: "wms", url: 'fakeURL'};
         let actions = {
-            onLayerAdd: () => {
-
+            onLayerAdd: (layer, options) => {
+                try {
+                    expect(layer.catalogURL).toBeFalsy();
+                    expect(layer.format).toBeTruthy();
+                    expect(layer.format).toBe(service.format);
+                    expect(options.zoomToLayer).toBeTruthy();
+                } catch (e) {
+                    done(e);
+                }
+                done();
             }
         };
-        const service = { format: "image/jpeg", title: "GeoServer WMS", type: "wms", url: 'fakeURL'};
-        let actionsSpy = expect.spyOn(actions, "onLayerAdd");
         const item = ReactDOM.render(<RecordItem
             record={sampleRecord}
             service={service}
@@ -302,80 +328,70 @@ describe('This test for RecordItem', () => {
             catalogURL="fakeURL"
             catalogType="wms"
         />, document.getElementById("container"));
-        expect(item).toExist();
+        expect(item).toBeTruthy();
 
         const itemDom = ReactDOM.findDOMNode(item);
-        expect(itemDom).toExist();
+        expect(itemDom).toBeTruthy();
         let button = TestUtils.findRenderedDOMComponentWithTag(
             item, 'button'
         );
-        expect(button).toExist();
+        expect(button).toBeTruthy();
         button.click();
-        expect(actionsSpy.calls.length).toBe(1);
-        expect(actionsSpy.calls[0].arguments.length).toBe(2);
-        expect(actionsSpy.calls[0].arguments[0].catalogURL).toNotExist();
-        expect(actionsSpy.calls[0].arguments[0].format).toBeTruthy();
-        expect(actionsSpy.calls[0].arguments[0].format).toBe(service.format);
-        expect(actionsSpy.calls[0].arguments[1].zoomToLayer).toBeTruthy();
     });
-    it('check wms default format', () => {
+    it('check wms default format', (done) => {
         let actions = {
-            onLayerAdd: () => {
-
+            onLayerAdd: (layer, options) => {
+                expect(layer.catalogURL).toBeFalsy();
+                expect(layer.format).toBe('image/png');
+                expect(options.zoomToLayer).toBeTruthy();
+                done();
             }
         };
-        let actionsSpy = expect.spyOn(actions, "onLayerAdd");
         const item = ReactDOM.render(<RecordItem
             record={sampleRecord}
             onLayerAdd={actions.onLayerAdd}
             catalogURL="fakeURL"
             catalogType="wms"
         />, document.getElementById("container"));
-        expect(item).toExist();
+        expect(item).toBeTruthy();
 
         const itemDom = ReactDOM.findDOMNode(item);
-        expect(itemDom).toExist();
+        expect(itemDom).toBeTruthy();
         let button = TestUtils.findRenderedDOMComponentWithTag(
             item, 'button'
         );
-        expect(button).toExist();
+        expect(button).toBeTruthy();
         button.click();
-        expect(actionsSpy.calls.length).toBe(1);
-        expect(actionsSpy.calls[0].arguments.length).toBe(2);
-        expect(actionsSpy.calls[0].arguments[0].catalogURL).toNotExist();
-        expect(actionsSpy.calls[0].arguments[0].format).toBe('image/png');
-        expect(actionsSpy.calls[0].arguments[1].zoomToLayer).toBeTruthy();
     });
-    it('check wms NOT default format', () => {
+    it('check wms NOT default format', (done) => {
         let actions = {
-            onLayerAdd: () => {
-
+            onLayerAdd: (layer, options) => {
+                expect(layer.catalogURL).toBeFalsy();
+                expect(layer.format).toBe('image/jpeg');
+                expect(options.zoomToLayer).toBeTruthy();
+                done();
             }
         };
-        let actionsSpy = expect.spyOn(actions, "onLayerAdd");
         const item = ReactDOM.render(<RecordItem
             record={{...sampleRecord, formats: ["image/jpeg"]}}
             onLayerAdd={actions.onLayerAdd}
             catalogURL="fakeURL"
             catalogType="wms"
         />, document.getElementById("container"));
-        expect(item).toExist();
+        expect(item).toBeTruthy();
 
         const itemDom = ReactDOM.findDOMNode(item);
-        expect(itemDom).toExist();
+        expect(itemDom).toBeTruthy();
         let button = TestUtils.findRenderedDOMComponentWithTag(
             item, 'button'
         );
-        expect(button).toExist();
+        expect(button).toBeTruthy();
         button.click();
-        expect(actionsSpy.calls.length).toBe(1);
-        expect(actionsSpy.calls[0].arguments.length).toBe(2);
-        expect(actionsSpy.calls[0].arguments[0].catalogURL).toNotExist();
-        expect(actionsSpy.calls[0].arguments[0].format).toBe('image/jpeg');
-        expect(actionsSpy.calls[0].arguments[1].zoomToLayer).toBeTruthy();
     });
-    it('check auth params to be removed (WMS)', () => {
+    it('check auth params to be removed (WMS)', (done) => {
         const recordToClean = {
+            serviceType: 'wms',
+            isValid: true,
             identifier: "test-identifier",
             title: "sample title",
             tags: ["subject1", "subject2"],
@@ -394,14 +410,24 @@ describe('This test for RecordItem', () => {
                 url: "http://wms.sample.service:80/geoserver/wms?SERVICE=WMS&ms2-authkey=TEST&requiredParam=REQUIRED",
                 SRS: [],
                 params: { name: "workspace:layername" }
-            }]
-        };
-        let actions = {
-            onLayerAdd: () => {
-
+            }],
+            ogcReferences: {
+                type: "OGC:WMS",
+                url: "http://wms.sample.service:80/geoserver/wms?SERVICE=WMS&ms2-authkey=TEST&requiredParam=REQUIRED",
+                SRS: [],
+                params: { name: "workspace:layername" }
             }
         };
-        let actionsSpy = expect.spyOn(actions, "onLayerAdd");
+        let actions = {
+            onLayerAdd: (layer, options) => {
+                expect(layer.catalogURL).toBeFalsy();
+                expect(layer.params).toBeTruthy();
+                expect(layer.params.requiredParam).toBe("REQUIRED");
+                expect(options.zoomToLayer).toBeTruthy();
+                expect(layer.params["ms2-authkey"]).toBeFalsy("auth param is passed in params list but it shouldn't");
+                done();
+            }
+        };
         const item = ReactDOM.render(<RecordItem
             authkeyParamNames={["ms2-authkey"]}
             record={recordToClean}
@@ -409,25 +435,20 @@ describe('This test for RecordItem', () => {
             catalogURL="fakeURL"
             catalogType="wms"
         />, document.getElementById("container"));
-        expect(item).toExist();
+        expect(item).toBeTruthy();
 
         const itemDom = ReactDOM.findDOMNode(item);
-        expect(itemDom).toExist();
+        expect(itemDom).toBeTruthy();
         let button = TestUtils.findRenderedDOMComponentWithTag(
             item, 'button'
         );
-        expect(button).toExist();
+        expect(button).toBeTruthy();
         button.click();
-        expect(actionsSpy.calls.length).toBe(1);
-        expect(actionsSpy.calls[0].arguments.length).toBe(2);
-        expect(actionsSpy.calls[0].arguments[0].catalogURL).toNotExist();
-        expect(actionsSpy.calls[0].arguments[0].params).toExist();
-        expect(actionsSpy.calls[0].arguments[0].params.requiredParam).toBe("REQUIRED");
-        expect(actionsSpy.calls[0].arguments[1].zoomToLayer).toBeTruthy();
-        expect(actionsSpy.calls[0].arguments[0].params["ms2-authkey"]).toNotExist("auth param is passed in params list but it shouldn't");
     });
-    it('check auth params to be removed (WMTS)', () => {
+    it('check auth params to be removed (WMTS)', (done) => {
         const recordToClean = {
+            serviceType: 'wmts',
+            isValid: true,
             identifier: "test-identifier",
             title: "sample title",
             tags: ["subject1", "subject2"],
@@ -446,14 +467,24 @@ describe('This test for RecordItem', () => {
                 url: "http://wms.sample.service:80/geoserver/gwc/service/wmts?ms2-authkey=TEST&requiredParam=REQUIRED",
                 SRS: ['EPSG:4326', 'EPSG:3857'],
                 params: { name: "workspace:layername" }
-            }]
-        };
-        let actions = {
-            onLayerAdd: () => {
-
+            }],
+            ogcReferences: {
+                type: "OGC:WMTS",
+                url: "http://wms.sample.service:80/geoserver/gwc/service/wmts?ms2-authkey=TEST&requiredParam=REQUIRED",
+                SRS: ['EPSG:4326', 'EPSG:3857'],
+                params: { name: "workspace:layername" }
             }
         };
-        let actionsSpy = expect.spyOn(actions, "onLayerAdd");
+        let actions = {
+            onLayerAdd: (layer, options) => {
+                expect(options.zoomToLayer).toBeTruthy();
+                expect(layer.catalogURL).toBeFalsy();
+                expect(layer.params).toBeTruthy();
+                expect(layer.params.requiredParam).toBe("REQUIRED");
+                expect(layer.params["ms2-authkey"]).toBeFalsy("auth param is passed in params list but it shouldn't");
+                done();
+            }
+        };
         const item = ReactDOM.render(<RecordItem
             authkeyParamNames={["ms2-authkey"]}
             record={recordToClean}
@@ -461,31 +492,25 @@ describe('This test for RecordItem', () => {
             catalogURL="fakeURL"
             catalogType="wms"
         />, document.getElementById("container"));
-        expect(item).toExist();
+        expect(item).toBeTruthy();
 
         const itemDom = ReactDOM.findDOMNode(item);
-        expect(itemDom).toExist();
+        expect(itemDom).toBeTruthy();
         let button = TestUtils.findRenderedDOMComponentWithTag(
             item, 'button'
         );
-        expect(button).toExist();
+        expect(button).toBeTruthy();
         button.click();
-        expect(actionsSpy.calls.length).toBe(1);
-        expect(actionsSpy.calls[0].arguments.length).toBe(2);
-        expect(actionsSpy.calls[0].arguments[1].zoomToLayer).toBeTruthy();
-        expect(actionsSpy.calls[0].arguments[0].catalogURL).toNotExist();
-        expect(actionsSpy.calls[0].arguments[0].params).toExist();
-        expect(actionsSpy.calls[0].arguments[0].params.requiredParam).toBe("REQUIRED");
-        expect(actionsSpy.calls[0].arguments[0].params["ms2-authkey"]).toNotExist("auth param is passed in params list but it shouldn't");
     });
 
-    it('check event handlers with layerBaseConfig and csw service', () => {
+    it('check event handlers with layerBaseConfig and csw service', (done) => {
         let actions = {
-            onLayerAdd: () => {
-
+            onLayerAdd: (layer, options) => {
+                expect(options.zoomToLayer).toBeTruthy();
+                expect(layer.extraProp).toBe('val1');
+                done();
             }
         };
-        let actionsSpy = expect.spyOn(actions, "onLayerAdd");
         const item = ReactDOM.render(<RecordItem
             record={sampleRecord}
             onLayerAdd={actions.onLayerAdd}
@@ -495,28 +520,25 @@ describe('This test for RecordItem', () => {
                 extraProp: 'val1'
             }}
         />, document.getElementById("container"));
-        expect(item).toExist();
+        expect(item).toBeTruthy();
 
         const itemDom = ReactDOM.findDOMNode(item);
-        expect(itemDom).toExist();
+        expect(itemDom).toBeTruthy();
         let button = TestUtils.findRenderedDOMComponentWithTag(
             item, 'button'
         );
-        expect(button).toExist();
+        expect(button).toBeTruthy();
         button.click();
-        expect(actionsSpy.calls.length).toBe(1);
-        expect(actionsSpy.calls[0].arguments.length).toBe(2);
-        expect(actionsSpy.calls[0].arguments[1].zoomToLayer).toBeTruthy();
-        expect(actionsSpy.calls[0].arguments[0].extraProp).toBe('val1');
     });
 
-    it('check event handlers with layerBaseConfig and wms service', () => {
+    it('check event handlers with layerBaseConfig and wms service', (done) => {
         let actions = {
-            onLayerAdd: () => {
-
+            onLayerAdd: (layer, options) => {
+                expect(options.zoomToLayer).toBeTruthy();
+                expect(layer.extraProp).toBe('val1');
+                done();
             }
         };
-        let actionsSpy = expect.spyOn(actions, "onLayerAdd");
         const item = ReactDOM.render(<RecordItem
             record={sampleRecord}
             onLayerAdd={actions.onLayerAdd}
@@ -526,19 +548,15 @@ describe('This test for RecordItem', () => {
                 extraProp: 'val1'
             }}
         />, document.getElementById("container"));
-        expect(item).toExist();
+        expect(item).toBeTruthy();
 
         const itemDom = ReactDOM.findDOMNode(item);
-        expect(itemDom).toExist();
+        expect(itemDom).toBeTruthy();
         let button = TestUtils.findRenderedDOMComponentWithTag(
             item, 'button'
         );
-        expect(button).toExist();
+        expect(button).toBeTruthy();
         button.click();
-        expect(actionsSpy.calls.length).toBe(1);
-        expect(actionsSpy.calls[0].arguments.length).toBe(2);
-        expect(actionsSpy.calls[0].arguments[1].zoomToLayer).toBeTruthy();
-        expect(actionsSpy.calls[0].arguments[0].extraProp).toBe('val1');
     });
 
     it('test create record item with no get capabilities links', () => {
@@ -546,9 +564,9 @@ describe('This test for RecordItem', () => {
         const component = ReactDOM.render(<RecordItem record={sampleRecord} showGetCapLinks/>,
             document.getElementById("container"));
         // check that the component was intanciated
-        expect(component).toExist();
+        expect(component).toBeTruthy();
         const componentDom = ReactDOM.findDOMNode(component);
-        expect(componentDom).toExist();
+        expect(componentDom).toBeTruthy();
         // we should have two buttons enable
         const buttons = componentDom.getElementsByTagName('button');
         expect(buttons.length).toBe(1);
@@ -559,9 +577,9 @@ describe('This test for RecordItem', () => {
         const component = ReactDOM.render(<RecordItem record={getCapRecord} showGetCapLinks/>,
             document.getElementById("container"));
         // check that the component was intanciated
-        expect(component).toExist();
+        expect(component).toBeTruthy();
         const componentDom = ReactDOM.findDOMNode(component);
-        expect(componentDom).toExist();
+        expect(componentDom).toBeTruthy();
         // we should have two buttons enable
         const buttons = componentDom.getElementsByTagName('button');
         expect(buttons.length).toBe(2);
@@ -572,9 +590,9 @@ describe('This test for RecordItem', () => {
         const component = ReactDOM.render(<RecordItem showGetCapLinks={false} record={getCapRecord}/>,
             document.getElementById("container"));
         // check that the component was intanciated
-        expect(component).toExist();
+        expect(component).toBeTruthy();
         const componentDom = ReactDOM.findDOMNode(component);
-        expect(componentDom).toExist();
+        expect(componentDom).toBeTruthy();
         // we should have only one button
         const buttons = componentDom.getElementsByTagName('button');
         expect(buttons.length).toBe(1);
@@ -591,39 +609,37 @@ describe('This test for RecordItem', () => {
             record={sampleRecord2}
             onError={actions.onError}
             crs="EPSG:3857"/>, document.getElementById("container"));
-        expect(item).toExist();
+        expect(item).toBeTruthy();
 
         let button = TestUtils.findRenderedDOMComponentWithTag(
             item, 'button'
         );
-        expect(button).toExist();
+        expect(button).toBeTruthy();
         button.click();
         expect(actionsSpy.calls.length).toBe(1);
     });
-    it('check add layer with bounding box', () => {
+    it('check add layer with bounding box', (done) => {
         let actions = {
-            onLayerAdd: () => {
-
+            onLayerAdd: (layer) => {
+                expect(layer.bbox).toBeTruthy();
+                expect(layer.bbox.crs).toBeTruthy();
+                expect(layer.bbox.bounds).toBeTruthy();
+                done();
             }
         };
-        let actionsSpy = expect.spyOn(actions, "onLayerAdd");
         const item = ReactDOM.render(<RecordItem
             record={sampleRecord}
             onLayerAdd={actions.onLayerAdd}
         />, document.getElementById("container"));
-        expect(item).toExist();
+        expect(item).toBeTruthy();
 
         const itemDom = ReactDOM.findDOMNode(item);
-        expect(itemDom).toExist();
+        expect(itemDom).toBeTruthy();
         let button = TestUtils.findRenderedDOMComponentWithTag(
             item, 'button'
         );
-        expect(button).toExist();
+        expect(button).toBeTruthy();
         button.click();
-        expect(actionsSpy.calls.length).toBe(1);
-        expect(actionsSpy.calls[0].arguments[0].bbox).toExist();
-        expect(actionsSpy.calls[0].arguments[0].bbox.crs).toExist();
-        expect(actionsSpy.calls[0].arguments[0].bbox.bounds).toExist();
     });
 
     // test title localization
@@ -654,10 +670,10 @@ describe('This test for RecordItem', () => {
     it('uses the correct localization', () => {
         // Default localization is en-US
         const item = ReactDOM.render(<RecordItem record={localizedRecord}/>, document.getElementById("container"));
-        expect(item).toExist();
+        expect(item).toBeTruthy();
 
         const itemDom = ReactDOM.findDOMNode(item);
-        expect(itemDom).toExist();
+        expect(itemDom).toBeTruthy();
         const identifiers = itemDom.getElementsByClassName('identifier');
         expect(identifiers.length).toBe(1);
         const titles = itemDom.getElementsByClassName('mapstore-side-card-title');
@@ -668,10 +684,10 @@ describe('This test for RecordItem', () => {
     it('uses the default localization', () => {
         // Default localization is en-US
         const item = ReactDOM.render(<RecordItem record={localizedRecord} currentLocale="it-IT"/>, document.getElementById("container"));
-        expect(item).toExist();
+        expect(item).toBeTruthy();
 
         const itemDom = ReactDOM.findDOMNode(item);
-        expect(itemDom).toExist();
+        expect(itemDom).toBeTruthy();
         const identifiers = itemDom.getElementsByClassName('identifier');
         expect(identifiers.length).toBe(1);
         const titles = itemDom.getElementsByClassName('mapstore-side-card-title');
@@ -682,10 +698,10 @@ describe('This test for RecordItem', () => {
     it('has not title', () => {
         // Default localization is en-US
         const item = ReactDOM.render(<RecordItem record={noTitleRecord} currentLocale="it-IT"/>, document.getElementById("container"));
-        expect(item).toExist();
+        expect(item).toBeTruthy();
 
         const itemDom = ReactDOM.findDOMNode(item);
-        expect(itemDom).toExist();
+        expect(itemDom).toBeTruthy();
         const identifiers = itemDom.getElementsByClassName('identifier');
         expect(identifiers.length).toBe(1);
         const titles = itemDom.getElementsByClassName('mapstore-side-card-title');
@@ -710,10 +726,10 @@ describe('This test for RecordItem', () => {
         };
 
         const item = ReactDOM.render(<RecordItem record={recordWithoutRef}/>, document.getElementById("container"));
-        expect(item).toExist();
+        expect(item).toBeTruthy();
 
         const itemDom = ReactDOM.findDOMNode(item);
-        expect(itemDom).toExist();
+        expect(itemDom).toBeTruthy();
         const dangerText = itemDom.getElementsByClassName('text-danger');
         expect(dangerText.length).toBe(1);
     });
@@ -721,18 +737,18 @@ describe('This test for RecordItem', () => {
     it('hide/show thumbnail', () => {
 
         let item = ReactDOM.render(<RecordItem record={longDescriptioRecord} hideThumbnail/>, document.getElementById("container"));
-        expect(item).toExist();
+        expect(item).toBeTruthy();
 
         let itemDom = ReactDOM.findDOMNode(item);
-        expect(itemDom).toExist();
+        expect(itemDom).toBeTruthy();
         let image = itemDom.getElementsByTagName('img');
         expect(image.length).toBe(0);
 
         item = ReactDOM.render(<RecordItem record={longDescriptioRecord} hideThumbnail={false}/>, document.getElementById("container"));
-        expect(item).toExist();
+        expect(item).toBeTruthy();
 
         itemDom = ReactDOM.findDOMNode(item);
-        expect(itemDom).toExist();
+        expect(itemDom).toBeTruthy();
         image = itemDom.getElementsByTagName('img');
         expect(image.length).toBe(1);
 
@@ -741,19 +757,19 @@ describe('This test for RecordItem', () => {
     it('hide/show identifier', () => {
 
         let item = ReactDOM.render(<RecordItem record={longDescriptioRecord} hideIdentifier/>, document.getElementById("container"));
-        expect(item).toExist();
+        expect(item).toBeTruthy();
 
         let itemDom = ReactDOM.findDOMNode(item);
-        expect(itemDom).toExist();
+        expect(itemDom).toBeTruthy();
         let identifier = itemDom.getElementsByClassName('identifier');
         expect(identifier.length).toBe(0);
 
 
         item = ReactDOM.render(<RecordItem record={longDescriptioRecord} hideIdentifier={false}/>, document.getElementById("container"));
-        expect(item).toExist();
+        expect(item).toBeTruthy();
 
         itemDom = ReactDOM.findDOMNode(item);
-        expect(itemDom).toExist();
+        expect(itemDom).toBeTruthy();
         identifier = itemDom.getElementsByClassName('identifier');
         expect(identifier.length).toBe(1);
 
@@ -761,22 +777,22 @@ describe('This test for RecordItem', () => {
 
     it('show Expand button', () => {
         let item = ReactDOM.render(<RecordItem record={longDescriptioRecord} hideExpand={false}/>, document.getElementById("container"));
-        expect(item).toExist();
+        expect(item).toBeTruthy();
 
         item = ReactDOM.render(<RecordItem record={longDescriptioRecord} hideExpand={false}/>, document.getElementById("container"));
         let itemDom = ReactDOM.findDOMNode(item);
-        expect(itemDom).toExist();
+        expect(itemDom).toBeTruthy();
         let expand = itemDom.getElementsByClassName('glyphicon-chevron-left');
         expect(expand.length).toBe(1);
 
     });
     it('hide Expand button', () => {
         let item = ReactDOM.render(<RecordItem record={longDescriptioRecord} hideExpand/>, document.getElementById("container"));
-        expect(item).toExist();
+        expect(item).toBeTruthy();
         let itemDom = ReactDOM.findDOMNode(item);
 
         itemDom = ReactDOM.findDOMNode(item);
-        expect(itemDom).toExist();
+        expect(itemDom).toBeTruthy();
         let expand = itemDom.getElementsByClassName('glyphicon-chevron-left');
         expect(expand.length).toBe(0);
     });
@@ -795,10 +811,10 @@ describe('This test for RecordItem', () => {
                 showTemplate
                 hideExpand={false}
             />, document.getElementById("container"));
-        expect(item).toExist();
+        expect(item).toBeTruthy();
 
         let itemDom = ReactDOM.findDOMNode(item);
-        expect(itemDom).toExist();
+        expect(itemDom).toBeTruthy();
         let expand = itemDom.getElementsByClassName('glyphicon-chevron-left');
         expect(expand.length).toBe(1);
         TestUtils.Simulate.click(expand[0]);
@@ -819,10 +835,10 @@ describe('This test for RecordItem', () => {
                 showTemplate
                 hideExpand={false}
             />, document.getElementById("container"));
-        expect(item).toExist();
+        expect(item).toBeTruthy();
 
         let itemDom = ReactDOM.findDOMNode(item);
-        expect(itemDom).toExist();
+        expect(itemDom).toBeTruthy();
         let expand = itemDom.getElementsByClassName('glyphicon-chevron-left');
         expect(expand.length).toBe(1);
         TestUtils.Simulate.click(expand[0]);
@@ -830,64 +846,65 @@ describe('This test for RecordItem', () => {
         expect(desc.length).toBe(1);
         expect(desc[0].innerText.indexOf("sample title and description catalog.notAvailable") !== -1).toBe(true);
     });
-    it('check default format is added to layer props', () => {
+    it('check default format is added to layer props', (done) => {
         const defaultFormat = 'image/jpeg';
         let actions = {
-            onLayerAdd: () => {}
+            onLayerAdd: (layer) => {
+                expect(layer.format).toBe(defaultFormat);
+                done();
+            }
         };
-        let actionsSpy = expect.spyOn(actions, "onLayerAdd");
         const item = ReactDOM.render(<RecordItem
             defaultFormat={defaultFormat}
             record={sampleRecord}
             onLayerAdd={actions.onLayerAdd}/>, document.getElementById("container"));
-        expect(item).toExist();
+        expect(item).toBeTruthy();
 
         const itemDom = ReactDOM.findDOMNode(item);
-        expect(itemDom).toExist();
+        expect(itemDom).toBeTruthy();
         let button = TestUtils.findRenderedDOMComponentWithTag(
             item, 'button'
         );
-        expect(button).toExist();
+        expect(button).toBeTruthy();
         button.click();
-        expect(actionsSpy.calls.length).toBe(1);
-        expect(actionsSpy.calls[0].arguments[0].format).toBe(defaultFormat);
     });
     it('check description is wrapped in span', () => {
         ReactDOM.render(<RecordItem record={sampleRecord}/>, document.getElementById("container"));
 
         const recordItemDiv = document.getElementsByClassName('mapstore-side-card')[0];
-        expect(recordItemDiv).toExist();
+        expect(recordItemDiv).toBeTruthy();
         const descDiv = recordItemDiv.getElementsByClassName('mapstore-side-card-desc')[0];
-        expect(descDiv).toExist();
+        expect(descDiv).toBeTruthy();
         expect(descDiv.hasChildNodes()).toBe(true);
         expect(descDiv.firstElementChild.tagName).toBe('SPAN');
     });
-    it('check formats are added to layer props (WMS)', () => {
+    it('check formats are added to layer props (WMS)', (done) => {
         const defaultFormat = 'image/jpeg';
         const formatOptions = ["image/png", "image/png8"];
         const infoFormatOptions = ["text/html", "text/plain"];
         let actions = {
-            onLayerAdd: () => {}
+            onLayerAdd: (layer) => {
+                expect(layer.format).toBe(defaultFormat);
+                expect(layer.imageFormats).toEqual(formatOptions);
+                expect(layer.infoFormats).toBe(infoFormatOptions);
+                done();
+            }
         };
-        let actionsSpy = expect.spyOn(actions, "onLayerAdd");
+
         const item = ReactDOM.render(<RecordItem
             defaultFormat={defaultFormat}
             formatOptions={formatOptions}
             infoFormatOptions={infoFormatOptions}
             record={sampleRecord}
             onLayerAdd={actions.onLayerAdd}/>, document.getElementById("container"));
-        expect(item).toExist();
+        expect(item).toBeTruthy();
 
         const itemDom = ReactDOM.findDOMNode(item);
-        expect(itemDom).toExist();
+        expect(itemDom).toBeTruthy();
         let button = TestUtils.findRenderedDOMComponentWithTag(
             item, 'button'
         );
-        expect(button).toExist();
+        expect(button).toBeTruthy();
         button.click();
-        expect(actionsSpy.calls.length).toBe(1);
-        expect(actionsSpy.calls[0].arguments[0].format).toBe(defaultFormat);
-        expect(actionsSpy.calls[0].arguments[0].imageFormats).toEqual(formatOptions);
-        expect(actionsSpy.calls[0].arguments[0].infoFormats).toBe(infoFormatOptions);
     });
 });
