+++ web/client/components/catalog/editor/AdvancedSettings/RasterAdvancedSettings.js
@@ -8,7 +8,7 @@
 import React, {useEffect} from 'react';
 import {FormGroup, Col, ControlLabel, Checkbox} from "react-bootstrap";
 import RS from 'react-select';
-import { DEFAULT_FORMAT_WMS } from '../../../../utils/CatalogUtils';
+import { DEFAULT_FORMAT_WMS } from '../../../../api/WMS';
 import localizedProps from '../../../misc/enhancers/localizedProps';
 const Select = localizedProps('noResultsText')(RS);
