+++ web/client/components/catalog/editor/MainFormUtils.js
@@ -2,11 +2,12 @@ import url from 'url';
 
 export const defaultPlaceholder = (service) => {
     let urlPlaceholder = {
-        wfs: "e.g. https://mydomain.com/geoserver/wfs",
-        wmts: "e.g. https://mydomain.com/geoserver/gwc/service/wmts",
-        wms: "e.g. https://mydomain.com/geoserver/wms",
-        csw: "e.g. https://mydomain.com/geoserver/csw",
-        tms: "e.g. https://mydomain.com/geoserver/gwc/service/tms/1.0.0"
+        "wfs": "e.g. https://mydomain.com/geoserver/wfs",
+        "wmts": "e.g. https://mydomain.com/geoserver/gwc/service/wmts",
+        "wms": "e.g. https://mydomain.com/geoserver/wms",
+        "csw": "e.g. https://mydomain.com/geoserver/csw",
+        "tms": "e.g. https://mydomain.com/geoserver/gwc/service/tms/1.0.0",
+        "3dtiles": "e.g. https://mydomain.com/tileset.json"
     };
     for ( const [key, value] of Object.entries(urlPlaceholder)) {
         if ( key === service.type) {
