+++ web/client/selectors/catalog.js
@@ -10,7 +10,7 @@ import { createSelector, createStructuredSelector } from 'reselect';
 import { get } from 'lodash';
 
 import { projectionSelector } from './map';
-import { DEFAULT_FORMAT_WMS, getUniqueInfoFormats } from "../utils/CatalogUtils";
+import { DEFAULT_FORMAT_WMS, getUniqueInfoFormats } from '../api/WMS';
 
 export const staticServicesSelector = (state) => get(state, "catalog.default.staticServices");
 export const servicesSelector = (state) => get(state, "catalog.services");
