+++ web/client/utils/LayersUtils.js
@@ -337,11 +337,10 @@ export const getDimension = (dimensions, dimension) => {
  * the layer id will returned will be something like `layerName__2` when 2 is the layer size (for retro compatibility, it should be removed in the future).
  * Otherwise a random string will be appended to the layer name.
  * @param {object} layer the layer
- * @param {array} [layers] an array to use to generate the id @deprecated
  * @returns {string} the id of the layer, or a generated one
  */
-export const getLayerId = (layerObj, layers) => {
-    return layerObj && layerObj.id || layerObj.name + "__" + (layers ? layers.length : Math.random().toString(36).substring(2, 15));
+export const getLayerId = (layerObj) => {
+    return layerObj && layerObj.id || `${layerObj.name ? `${layerObj.name}__` : ''}${uuidv1()}`;
 };
 /**
  * Normalizes the layer to assign missing Ids
